design simplification.
using a plain simple design to check if a step is needed or not.
We check for a result to be present in the context object: if it is null we run the step.
Steps are fixed and against OCP and SRP if we need to add a step we need to change two objects adding functions.